Merging changes from master into branch

+++ flask-api/flaskr/party.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, abort, jsonify, request, make_response
 from flaskr.db import get_db
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 import json
 
 bp = Blueprint('parties', __name__, url_prefix='/parties')
@@ -8,6 +8,11 @@ bp = Blueprint('parties', __name__, url_prefix='/parties')
 @bp.route('/create', methods=['POST'])
 @jwt_required
 def create():
+    # Check Authorisation
+    token = get_jwt_claims()
+    if token['user_type'] != 'commissioner':
+        abort(403, 'Forbidden')
+
     if not(request.data):
         abort(400, 'Bad Request')
 
@@ -18,9 +23,10 @@ def create():
         abort(400, 'Bad Request')
 
     db = get_db()
+    cursor = db.cursor()
 
     # Check if the Voting Event exists
-    v_event = db.execute(
+    v_event = cursor.execute(
         'SELECT * FROM voting_event WHERE id = ?', (v_event_id,)
     ).fetchone()
 
@@ -28,13 +34,15 @@ def create():
         abort(400, 'Bad Request')
 
     # Save to DB
-    db.execute(
+    cursor.execute(
         'INSERT INTO party (party_name, v_event_id) VALUES (?, ?)',
         (party_name, v_event_id)
     )
     db.commit()
 
-    return '', 204
+    resp = make_response({"id": cursor.lastrowid}, 201)
+    resp.headers['Content-Type'] = 'application/json'
+    return resp
 
 @bp.route('/<party_id>', methods=['GET'])
 @jwt_required
@@ -101,6 +109,11 @@ def getList():
 @bp.route('/<int:party_id>/update', methods=['PUT'])
 @jwt_required
 def update(party_id):
+    # Check Authorisation
+    token = get_jwt_claims()
+    if token['user_type'] != 'commissioner':
+        abort(403, 'Forbidden')
+
     if not(request.data):
         abort(400, 'Bad Request')
 
@@ -111,9 +124,10 @@ def update(party_id):
         abort(400, 'Bad Request')
 
     db = get_db()
+    cursor = db.cursor()
 
     # Check if the Voting Event exists
-    v_event = db.execute(
+    v_event = cursor.execute(
         'SELECT * FROM voting_event WHERE id = ?', (v_event_id,)
     ).fetchone()
 
@@ -121,7 +135,7 @@ def update(party_id):
         abort(400, 'Bad Request')
 
     # Update DB
-    db.execute(
+    cursor.execute(
         'UPDATE party SET party_name = ?, v_event_id = ? WHERE id = ?',
         (party_name, v_event_id, party_id)
     )
@@ -132,10 +146,16 @@ def update(party_id):
 @bp.route('/<int:party_id>/delete', methods=['DELETE'])
 @jwt_required
 def delete(party_id):
+    # Check Authorisation
+    token = get_jwt_claims()
+    if token['user_type'] != 'commissioner':
+        abort(403, 'Forbidden')
+
     db = get_db()
+    cursor = db.cursor()
 
     # Save to DB
-    db.execute(
+    cursor.execute(
         'DELETE FROM party WHERE id = ?', (party_id,)
     )
